Switch to standard types, replace long types
Includes `<cstdint>` and updates typedefs to point to it so that I don't
need to go around the entire source and update every single type.

Replace `long` with `int` in any place where the type is used for a
plain value. This clarifies that nothing is broken here.

Generally use `int` as the "don't care" type. `int` always means `int32_t`
on modern compilers on all platforms. This is convention.

`unsigned long` becomes `uint32_t`.

`printf` function which was converting pointers to int to print them
now uses `%p` instead to print the pointer directly.

`EffectInfo::Source` is a suspicious union of pointer and enum types.
I used `intptr_t` here for comparisons to prevent truncation.

Did not touch `long` in preprocessor or tokenizer sources.