Replace min/max macros with standard C++
The simplest solution was to add `using` to replace the macros exactly
with the C++ functions.

Platform toolset has been increased to v143.
This was probably not needed for this PR but it is needed for C++17 and libtcod.

An explicit type must be given when int constants are mixed with smaller types.

The sources which needed to be updated includes `Grammar.acc`.

Refactor Grammar.acc to use std:: namespace