Fix NULL Map dereference in MapIterate
Missing NULL test in the for-loop initializer meant that the `map` pointer
was accessed before being tested.

I've documented the macro and made it safer to use.
It might be possible to remove the `i` parameter since I don't think
anything using this macro accesses that parameter.
I'm not sure if `map` needs null testing every loop.